Move the deme expasion stuff to its own directory
It appears to sharre no commmon code or mechanisms with the
metapopulation, so move it out of the way.
