+++ opencog/learning/moses/deme/CMakeLists.txt
@@ -1,0 +1,10 @@
+#install header files
+INSTALL(FILES
+    deme_expander.h
+    deme_params.h
+    feature_selector.h
+
+	DESTINATION
+
+	"include/${PROJECT_NAME}/learning/moses/demes"
+)
