+++ opencog/learning/moses/example-progs/moses-ann-pole2-hillclimbing.cc
@@ -5,7 +5,7 @@
 #include <opencog/comboreduct/interpreter/eval.h>
 
 
-#include "../metapopulation/deme_expander.h"
+#include "../deme/deme_expander.h"
 #include "../metapopulation/metapopulation.h"
 
 #include "../moses/moses_main.h"
