+++ opencog/learning/moses/example-progs/moses-ann-pole2.cc
@@ -26,7 +26,7 @@
 
 #include <opencog/comboreduct/interpreter/eval.h>
 
-#include "../metapopulation/deme_expander.h"
+#include "../deme/deme_expander.h"
 #include "../metapopulation/metapopulation.h"
 
 #include "../representation/representation.h"
