+++ opencog/learning/moses/example-progs/moses-ann-pole2nv.cc
@@ -4,7 +4,7 @@
 #include <opencog/util/Logger.h>
 #include <opencog/comboreduct/interpreter/eval.h>
 
-#include "../metapopulation/deme_expander.h"
+#include "../deme/deme_expander.h"
 #include "../metapopulation/metapopulation.h"
 
 #include "../representation/representation.h"
