+++ opencog/learning/moses/example-progs/moses-ann-xor.cc
@@ -6,7 +6,7 @@
 
 #include <opencog/comboreduct/interpreter/eval.h>
 
-#include "../metapopulation/deme_expander.h"
+#include "../deme/deme_expander.h"
 #include "../metapopulation/metapopulation.h"
 
 #include "../moses/moses_main.h"
