+++ opencog/learning/moses/metapopulation/CMakeLists.txt
@@ -1,9 +1,7 @@
 #install header files
 INSTALL(FILES
-    deme_expander.h
     metapopulation.h
     metapop_params.h
-    feature_selector.h
 
 	DESTINATION
