+++ opencog/learning/moses/moses/local_moses.h
@@ -24,8 +24,8 @@
 #ifndef _MOSES_LOCAL_MOSES_H
 #define _MOSES_LOCAL_MOSES_H
 
+#include "../deme/deme_expander.h"
 #include "../metapopulation/metapopulation.h"
-#include "../metapopulation/deme_expander.h"
 #include "moses_params.h"
 
 namespace opencog {
