+++ opencog/learning/moses/moses/moses_main.cc
@@ -51,52 +51,6 @@ const char * version_string =
 
 #endif
 
-void print_stats_header (optim_stats *os, bool diversity)
-{
-    // Print legend for the columns of the stats.
-    if (logger().isInfoEnabled()) {
-        stringstream ss;
-        ss << "Stats: # \n"
-           << "Stats: # Stats are tab-separated, ready for graphing.\n"
-           << "Stats: # You can also use the script parse_log.py to extract a CSV file given a moses log file.\n"
-           << "Column explanation:\n"
-           << "Stats: # \n"
-           << "Stats: # gen is the generation number.\n"
-           << "Stats: # num_evals is the number of scoring function evaluations so far.\n"
-           << "Stats: # elapsed is the wall-clock time, in seconds, since start.\n"
-           << "Stats: # metapop_size is the size of the metapopulation.\n"
-           << "Stats: # best_score is the highest raw score seen, of all exemplars.\n"
-           << "Stats: # complexity is in bits, of the highest-composite score exemplar.\n";
-        if (os) {
-           ss << "Stats: # field_set_size is the ESTIMATED number of bits in all the knobs.\n"
-              << "Stats: # optim_steps is the number of steps the optimizer took.\n"
-              << "Stats: # over_budget is bool, T if search exceeded scoring func eval budget.\n";
-        }
-        if (diversity) {
-            ss << "Stats: # n_pairs is the number of pairs of candidates used to compute the diversity stats.\n"
-               << "Stats: # mean_dst is the average bscore distance between 2 candidates.\n"
-               << "Stats: # std_dst is the standard deviation of the average bscore distance between 2 candidates.\n"
-               << "Stats: # min_dst is the minimum bscore distance between 2 candidates.\n"
-               << "Stats: # max_dst is the maximum bscore distance between 2 candidates.\n";
-            ss << "Stats: # best_n_pairs is the number of pairs of candidates used to compute the diversity stats amongst the best candidates to be output.\n"
-               << "Stats: # best_mean_dst is the average bscore distance between 2 candidates amongst the best candidates to be output.\n"
-               << "Stats: # best_std_dst is the standard deviation of the average bscore distance between 2 candidates amongst the best candidates to be output.\n"
-               << "Stats: # best_min_dst is the minimum bscore distance between 2 candidates amongst the best candidates to be output.\n"
-               << "Stats: # best_max_dst is the maximum bscore distance between 2 candidates amongst the best candidates to be output.\n";
-        }
-        ss << "Stats: # \n"
-           << "Stats: # gen\tnum_evals\telapsed\tmetapop_size\tbest_score\tcomplexity";
-        if (os) {
-            ss << "\tfield_set_size\toptim_steps\tover_budget";
-        }
-        if (diversity) {
-            ss << "\tn_pairs\tmean_dst\tstd_dst\tmin_dst\tmax_dst"
-               << "\tbest_n_pairs\tbest_mean_dst\tbest_std_dst\tbest_min_dst\tbest_max_dst";                
-        }
-        logger().info(ss.str());
-    }
-}
-
 void run_moses(metapopulation& metapop,
                deme_expander& dex,
                const moses_parameters& moses_params,
