+++ tests/learning/moses/AntUTest.cxxtest
@@ -83,9 +83,11 @@ public:
 
         perceptions.insert(combo_tree(get_instance(id::is_food_ahead)));
 
+        deme_parameters demeparms;
+        demeparms.perceptions = &perceptions;
+        demeparms.actions = &actions;
+
         metapop_parameters metaparms;
-        metaparms.perceptions = &perceptions;
-        metaparms.actions = &actions;
         metaparms.complexity_temperature = 100;
 
         // Define optimization algo
@@ -96,15 +98,16 @@ public:
         hc_params.crossover = true;
         hill_climbing hc(opt_params, hc_params);
 
-        metapopulation metapop(combo_tree(id::sequential_and), tt,
-                               action_reduction(),
-                               cscorer, bscorer, hc, metaparms);
+        deme_expander dex(tt, action_reduction(),
+                              action_reduction(), cscorer, hc, demeparms);
+        metapopulation metapop(combo_tree(id::sequential_and),
+                               cscorer, bscorer, metaparms);
   
         boost::program_options::variables_map vm;
 
         moses_parameters moses_param(vm, jobs, true, max_evals, -1, 0, 100);
         moses_statistics st;
-        run_moses(metapop, moses_param, st);
+        run_moses(metapop, dex, moses_param, st);
         TS_ASSERT_EQUALS(metapop.best_score(), 0.0);
     }
 };
